# submissions/suy_tim_nhom27/api/openapi.yaml
openapi: 3.0.3
info:
  title: Nhóm 27 — API Gateway stub
  description: Tuần 3. Chỉ dữ liệu synthetic, lưu trong RAM. Không trả chẩn đoán hay
    đơn thuốc thực.
  version: 0.1.0
paths:
...
  /api/v1/auth/token:
    post:
      tags:
      - Identity
      summary: Login
      operationId: login_api_v1_auth_token_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Token thiếu/hết hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Không đúng vai trò hoặc quyền ca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Không tìm thấy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Xung đột phiên bản/trạng thái
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Chưa cấu hình demo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/cases:
    get:
      tags:
      - Cases
      summary: List Cases
      operationId: list_cases_api_v1_cases_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CaseOut'
                type: array
                title: Response List Cases Api V1 Cases Get
        '401':
          description: Token thiếu/hết hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Không đúng vai trò hoặc quyền ca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Không tìm thấy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Xung đột phiên bản/trạng thái
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Chưa cấu hình demo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - HTTPBearer: []
    post:
      tags:
      - Cases
      summary: Create Case
      operationId: create_case_api_v1_cases_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseCreateInput'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseOut'
        '401':
          description: Token thiếu/hết hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Không đúng vai trò hoặc quyền ca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Không tìm thấy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Xung đột phiên bản/trạng thái
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Chưa cấu hình demo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - HTTPBearer: []
  /api/v1/cases/{case_id}:
    get:
      tags:
      - Cases
      summary: Get Case
      operationId: get_case_api_v1_cases__case_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Case Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseOut'
        '401':
          description: Token thiếu/hết hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Không đúng vai trò hoặc quyền ca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Không tìm thấy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Xung đột phiên bản/trạng thái
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Chưa cấu hình demo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
      - Cases
      summary: Update Case
      operationId: update_case_api_v1_cases__case_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Case Id
      - name: expected_revision
        in: query
        required: true
        schema:
          type: integer
          title: Expected Revision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseInput'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseOut'
        '401':
          description: Token thiếu/hết hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Không đúng vai trò hoặc quyền ca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Không tìm thấy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Xung đột phiên bản/trạng thái
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Chưa cấu hình demo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/cases/{case_id}/encounters:
    post:
      tags:
      - Cases
      summary: Create Encounter
      operationId: create_encounter_api_v1_cases__case_id__encounters_post
      security:
      - HTTPBearer: []
      parameters:
      - name: case_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Case Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EncounterInput'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncounterOut'
        '401':
          description: Token thiếu/hết hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Không đúng vai trò hoặc quyền ca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Không tìm thấy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Xung đột phiên bản/trạng thái
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Chưa cấu hình demo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/encounters/{encounter_id}:
    get:
      tags:
      - Cases
      summary: Get Encounter
      operationId: get_encounter_api_v1_encounters__encounter_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: encounter_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Encounter Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EncounterOut'
        '401':
          description: Token thiếu/hết hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Không đúng vai trò hoặc quyền ca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Không tìm thấy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Xung đột phiên bản/trạng thái
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Chưa cấu hình demo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/encounters/{encounter_id}/observations:
    post:
      tags:
      - Clinical data
      summary: Observation
      operationId: observation_api_v1_encounters__encounter_id__observations_post
      security:
      - HTTPBearer: []
      parameters:
      - name: encounter_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Encounter Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationInput'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservationOut'
        '401':
          description: Token thiếu/hết hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Không đúng vai trò hoặc quyền ca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Không tìm thấy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Xung đột phiên bản/trạng thái
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Chưa cấu hình demo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
      - Clinical data
      summary: Observations
      operationId: observations_api_v1_encounters__encounter_id__observations_get
      security:
      - HTTPBearer: []
      parameters:
      - name: encounter_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Encounter Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ObservationOut'
                title: Response Observations Api V1 Encounters  Encounter Id  Observations
                  Get
        '401':
          description: Token thiếu/hết hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Không đúng vai trò hoặc quyền ca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Không tìm thấy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Xung đột phiên bản/trạng thái
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Chưa cấu hình demo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/encounters/{encounter_id}/medications:
    post:
      tags:
      - Clinical data
      summary: Medication
      operationId: medication_api_v1_encounters__encounter_id__medications_post
      security:
      - HTTPBearer: []
      parameters:
      - name: encounter_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Encounter Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicationInput'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicationOut'
        '401':
          description: Token thiếu/hết hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Không đúng vai trò hoặc quyền ca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Không tìm thấy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Xung đột phiên bản/trạng thái
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Chưa cấu hình demo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
      - Clinical data
      summary: Medications
      operationId: medications_api_v1_encounters__encounter_id__medications_get
      security:
      - HTTPBearer: []
      parameters:
      - name: encounter_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Encounter Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MedicationOut'
                title: Response Medications Api V1 Encounters  Encounter Id  Medications
                  Get
        '401':
          description: Token thiếu/hết hạn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Không đúng vai trò hoặc quyền ca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Không tìm thấy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Xung đột phiên bản/trạng thái
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Chưa cấu hình demo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    CaseCreateInput:
      properties:
        age:
          type: integer
          maximum: 120.0
          minimum: 0.0
          title: Age
        sex:
          type: string
          enum:
          - male
          - female
          - unknown
          title: Sex
        synthetic_code:
          title: Synthetic Code
          nullable: true
          type: string
          pattern: ^SYN-[A-Z0-9-]{1,24}$
      additionalProperties: false
      type: object
      required:
      - age
      - sex
      title: CaseCreateInput
    CaseInput:
      properties:
        synthetic_code:
          type: string
          pattern: ^SYN-[A-Z0-9-]{1,24}$
          title: Synthetic Code
        age:
          type: integer
          maximum: 120.0
          minimum: 0.0
          title: Age
        sex:
          type: string
          enum:
          - male
          - female
          - unknown
          title: Sex
      additionalProperties: false
      type: object
      required:
      - synthetic_code
      - age
      - sex
      title: CaseInput
    CaseOut:
      properties:
        synthetic_code:
          type: string
          pattern: ^SYN-[A-Z0-9-]{1,24}$
          title: Synthetic Code
        age:
          type: integer
          maximum: 120.0
          minimum: 0.0
          title: Age
        sex:
          type: string
          enum:
          - male
          - female
          - unknown
          title: Sex
        id:
          type: string
          format: uuid
          title: Id
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        revision:
          type: integer
          title: Revision
      additionalProperties: false
      type: object
      required:
      - synthetic_code
      - age
      - sex
      - id
      - owner_id
      - revision
      title: CaseOut
...
    Error:
      properties:
        detail:
          type: string
          title: Detail
      additionalProperties: false
      type: object
      required:
      - detail
      title: Error
...
    Login:
      properties:
        username:
          type: string
          enum:
          - doctor
          - pharmacist
          - admin
          title: Username
        password:
          type: string
          maxLength: 200
          minLength: 1
          title: Password
      additionalProperties: false
      type: object
      required:
      - username
      - password
      title: Login
...
    Token:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
          enum:
          - bearer
        expires_in:
          type: integer
          title: Expires In
          default: 3600
      additionalProperties: false
      type: object
      required:
      - access_token
      title: Token
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
      description: Opaque demo token from /api/v1/auth/token; expires after 3600 seconds.
        Production identity adapter pending.
servers:
- url: http://127.0.0.1:8000
  description: Local stub only; deployment requires HTTPS